requirements.txt
alembic==1.7.4
anyio==3.3.4
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.70.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.4.0
iniconfig==1.1.1
Mako==1.1.5
MarkupSafe==2.0.1
packaging==21.2
passlib==1.7.4
pluggy==1.0.0
psycopg2==2.9.1
py==1.11.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.5
python-dotenv==0.19.1
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.26
starlette==0.16.0
toml==0.10.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
zipp==3.6.0
